Eliminate redundant code in src
