+++ .gitignore
@@ -1,3 +1,10 @@
+# Data
+data/*
+!data/.gitkeep
+
+# Lightning logs
+/lightning_logs
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
@@ -158,7 +165,3 @@ cython_debug/
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
-
-# Data folders
-/data
-/lightning_logs
