+++ src/deep_learning/lstm.py
@@ -1,141 +1,7 @@
-# Torch classes & functions used in notebook 3.1, for the LSTM model.
-import pandas as pd
-import numpy as np
 import torch
 import lightning as L
 
-
-class sequence_scaler:
-    """
-    Takes in lists of dataframes, where each dataframe is an input or output sequence.
-
-    Returns scaled 3D numpy array of shape (observations, timesteps, features).
-    Can also backtransform scaled predictions.
-    """
-
-    def __init__(self, feature_range = (-1, 1)):
-        self.lower = feature_range[0]
-        self.upper = feature_range[1]
-
-    def fit(self, input_df, output_df):
-
-        # Get input & output sequences as 3D arrays
-        # The time index will be skipped, yielding shape (N, seq_length, seq_dims)
-        input = np.stack(input_df, axis = 0)
-        output = np.stack(output_df, axis = 0)
-
-        # Get number of dimensions
-        self.num_dimensions = input.shape[2]
-        
-        # Extract & save minimum, maximum for each dimension
-        dimensions_mini = []
-        dimensions_maxi = []
-        for dimension in range(0, self.num_dimensions):
-            min = np.min([
-                np.min(input[:, :, dimension]),
-                np.min(output[:, :, dimension])
-            ])
-            dimensions_mini.append(min)
-
-            max = np.max([
-                np.max(input[:, :, dimension]),
-                np.max(output[:, :, dimension])
-            ])
-            dimensions_maxi.append(max)
-
-        self.dimensions_mini = dimensions_mini
-        self.dimensions_maxi = dimensions_maxi
-
-    def transform(self, scale_df):
-
-        # Get sequence as 3D arrays
-        scale_array = np.stack(scale_df, axis = 0)
-
-        # Initialize list of scaled dimensions
-        scaled_dimensions = []
-
-        # Scale each dimension & append to list
-        for dimension in range(0, self.num_dimensions):
-            values = scale_array[:, :, dimension]
-            min = self.dimensions_mini[dimension]
-            max = self.dimensions_maxi[dimension]
-            std = (values - min) / (max - min)
-            scaled = std * (self.upper - self.lower) + self.lower
-            scaled_dimensions.append(scaled)
-
-        # Stack over 3rd axis & return
-        return np.stack(scaled_dimensions, axis = 2)
-
-    def backtransform_preds(self, preds_array, fitted_preds_dim = 0):
-
-        # Get n. of predicted quantiles to backtransform
-        n_quantiles = preds_array.shape[-1]
-
-        # Get the fitted mini & maxi for predictions
-        min = self.dimensions_mini[fitted_preds_dim] 
-        max = self.dimensions_maxi[fitted_preds_dim]
-
-        # Initialize list of backtransformed quantiles
-        backtrafo_quantiles = []
-
-        # Backtransform each quantile & append to list
-        for quantile in range(0, n_quantiles):
-            scaled = preds_array[:, :, quantile]
-            std = (scaled - self.lower) / (self.upper - self.lower)
-            values = std * (max - min) + min
-            backtrafo_quantiles.append(values)
-            
-        # Stack over 3rd axis & return
-        return np.stack(backtrafo_quantiles, axis = 2)
-
-
-class SequenceDataset(torch.utils.data.Dataset):
-    """
-    Simply takes in the input & output sequences as 3D arrays and returns them as Torch tensors.
-    """
-
-    # Store preprocessed input & output sequences
-    def __init__(self, input_seq, output_seq): 
-        self.input_seq = torch.tensor(input_seq, dtype = torch.float32) # Store input sequences
-        self.output_seq = torch.tensor(output_seq, dtype = torch.float32) # Store output sequences
-  
-    # Return data length  
-    def __len__(self):
-        return len(self.input_seq) 
-  
-    # Return a pair of input & output sequences
-    def __getitem__(self, idx):
-        return self.input_seq[idx], self.output_seq[idx]
-
-
-class QuantileLoss:
-    """
-    Takes in targets of shape (...),
-    predictions of shape (..., n_quantiles),
-    quantiles list.
-    
-    Returns unreduced quantile loss tensor of shape (..., n_quantiles),
-    where each value is quantile loss * 2 (equal to the MAE for q = 0.5).
-    
-    Implemented from pytorch_forecasting.metrics.quantile.QuantileLoss.
-    """
-
-    def __init__(self, quantiles):
-        self.quantiles = quantiles
-
-    def loss(self, pred, target):
-        
-        quantile_losses = []
-        for i, q in enumerate(self.quantiles):
-            error = target - pred[..., i]
-            quantile_error = torch.max(
-                (q - 1) * error,
-                q * error
-            ).unsqueeze(-1)
-            quantile_losses.append(quantile_error)
-
-        quantile_losses = torch.cat(quantile_losses, dim = 2)
-        return quantile_losses
+from src.deep_learning.quantile_loss import QuantileLoss
 
 
 class StatefulQuantileLSTM(L.LightningModule):
@@ -147,6 +13,9 @@ class StatefulQuantileLSTM(L.LightningModule):
 
     # Initialize model
     def __init__(self, hyperparams_dict):
+    """
+    hyperparams_dict: Dictionary of hyperparameter values. See below for required hyperparameters.
+    """
         
          # Delegate function to parent class
         super().__init__() 
@@ -155,14 +24,14 @@ class StatefulQuantileLSTM(L.LightningModule):
         self.save_hyperparameters(logger = False) 
 
         # Define hyperparameters
-        self.output_length = hyperparams_dict["output_length"] # Length of output sequence
-        self.input_size = hyperparams_dict["input_size"] # Number of features (network inputs)
-        self.horizon_start = hyperparams_dict["horizon_start"] # Start of the forecast horizon relevant for loss computing
-        self.quantiles = hyperparams_dict["quantiles"] # Provide as list of floats: [0.025, 0.5, 0.975]
+        self.output_length = hyperparams_dict["output_length"]  # Length of output sequence
+        self.input_size = hyperparams_dict["input_size"]  # Number of features (network inputs)
+        self.horizon_start = hyperparams_dict["horizon_start"]  # Start of the forecast horizon relevant for loss computing
+        self.quantiles = hyperparams_dict["quantiles"]  # Provide as list of floats: [0.025, 0.5, 0.975]
         self.learning_rate = hyperparams_dict["learning_rate"]
-        self.lr_decay = hyperparams_dict["lr_decay"]
-        self.num_layers = hyperparams_dict["num_layers"] # Number of layers in the LSTM block
-        self.hidden_size = hyperparams_dict["hidden_size"] # Number of units in each LSTM block = LSTM block output size
+        self.lr_decay = hyperparams_dict["lr_decay"]  # Learning rate decay, 0.9 to 1 in practice, 1 = no decay
+        self.num_layers = hyperparams_dict["num_layers"]  # Number of layers in the LSTM block
+        self.hidden_size = hyperparams_dict["hidden_size"]  # Number of units in each LSTM block = LSTM block output size
         self.dropout_rate = hyperparams_dict["dropout_rate"]
 
         # Define architecture
@@ -466,4 +335,3 @@ class StatefulQuantileLSTM(L.LightningModule):
           "scheduler": lr_scheduler
           }
         }
-
