+++ src/deep_learning/quantile_loss.py
@@ -1,0 +1,32 @@
+import torch
+
+
+class QuantileLoss:
+    """
+    Takes in targets of shape (...),
+    predictions of shape (..., n_quantiles),
+    quantiles list.
+    
+    Returns unreduced quantile loss tensor of shape (..., n_quantiles),
+    where each value is quantile loss * 2 (equal to the MAE for q = 0.5).
+    
+    Implemented from pytorch_forecasting.metrics.quantile.QuantileLoss.
+    """
+
+    def __init__(self, quantiles):
+        self.quantiles = quantiles
+
+    def loss(self, pred, target):
+        
+        quantile_losses = []
+        for i, q in enumerate(self.quantiles):
+            error = target - pred[..., i]
+            quantile_error = torch.max(
+                (q - 1) * error,
+                q * error
+            ).unsqueeze(-1)
+            quantile_losses.append(quantile_error)
+
+        quantile_losses = torch.cat(quantile_losses, dim = 2)
+        return quantile_losses
+
