+++ src/deep_learning/transformer.py
@@ -1,141 +1,7 @@
-# Torch classes & functions used in notebook 3.2, for the Transformer model.
-import pandas as pd
-import numpy as np
 import torch
 import lightning as L
 
-
-class sequence_scaler:
-    """
-    Takes in lists of dataframes where each dataframe is a source or target sequence.
-
-    Returns scaled 3D numpy arrays of shape (observations, timesteps, features).
-    Can also backtransform scaled predictions.
-    """
-
-    def __init__(self, feature_range = (-1, 1)):
-        self.lower = feature_range[0]
-        self.upper = feature_range[1]
-
-    def fit(self, source_df, target_df):
-
-        # Get source & target sequences as 3D arrays
-        # The time index will be skipped, yielding shape (N, seq_length, seq_dims)
-        source = np.stack(source_df, axis = 0)
-        target = np.stack(target_df, axis = 0)
-
-        # Get number of dimensions
-        self.num_dimensions = source.shape[2]
-        
-        # Extract & save minimum, maximum for each dimension
-        dimensions_mini = []
-        dimensions_maxi = []
-        for dimension in range(0, self.num_dimensions):
-            min = np.min([
-                np.min(source[:, :, dimension]),
-                np.min(target[:, :, dimension])
-            ])
-            dimensions_mini.append(min)
-
-            max = np.max([
-                np.max(source[:, :, dimension]),
-                np.max(target[:, :, dimension])
-            ])
-            dimensions_maxi.append(max)
-
-        self.dimensions_mini = dimensions_mini
-        self.dimensions_maxi = dimensions_maxi
-
-    def transform(self, scale_df):
-
-        # Get sequence as 3D arrays
-        scale_array = np.stack(scale_df, axis = 0)
-
-        # Initialize list of scaled dimensions
-        scaled_dimensions = []
-
-        # Scale each dimension & append to list
-        for dimension in range(0, self.num_dimensions):
-            values = scale_array[:, :, dimension]
-            min = self.dimensions_mini[dimension]
-            max = self.dimensions_maxi[dimension]
-            std = (values - min) / (max - min)
-            scaled = std * (self.upper - self.lower) + self.lower
-            scaled_dimensions.append(scaled)
-
-        # Stack over 3rd axis & return
-        return np.stack(scaled_dimensions, axis = 2)
-
-    def backtransform_preds(self, preds_array, fitted_preds_dim = 0):
-
-        # Get n. of predicted quantiles to backtransform
-        n_quantiles = preds_array.shape[-1]
-
-        # Get the fitted mini & maxi for predictions
-        min = self.dimensions_mini[fitted_preds_dim] 
-        max = self.dimensions_maxi[fitted_preds_dim]
-
-        # Initialize list of backtransformed quantiles
-        backtrafo_quantiles = []
-
-        # Backtransform each quantile & append to list
-        for quantile in range(0, n_quantiles):
-            scaled = preds_array[:, :, quantile]
-            std = (scaled - self.lower) / (self.upper - self.lower)
-            values = std * (max - min) + min
-            backtrafo_quantiles.append(values)
-            
-        # Stack over 3rd axis & return
-        return np.stack(backtrafo_quantiles, axis = 2)
-
-
-class SequenceDataset(torch.utils.data.Dataset):
-    """
-    Simply takes in the source & target sequences as 3D arrays and returns them as Torch tensors.
-    """
-
-    # Store preprocessed source & target sequences
-    def __init__(self, source_seq, target_seq): 
-        self.source_seq = torch.tensor(source_seq, dtype = torch.float32) # Store source sequences
-        self.target_seq = torch.tensor(target_seq, dtype = torch.float32) # Store target sequences
-  
-    # Return data length  
-    def __len__(self):
-        return len(self.source_seq) 
-  
-    # Return a pair of source & target sequences
-    def __getitem__(self, idx):
-        return self.source_seq[idx], self.target_seq[idx]
-
-
-class QuantileLoss:
-    """
-    Takes in targets of shape (...),
-    predictions of shape (..., n_quantiles),
-    quantiles list.
-    
-    Returns unreduced quantile loss tensor of shape (..., n_quantiles),
-    where each value is quantile loss * 2 (equal to the MAE for q = 0.5).
-    
-    Implemented from pytorch_forecasting.metrics.quantile.QuantileLoss.
-    """
-
-    def __init__(self, quantiles):
-        self.quantiles = quantiles
-
-    def loss(self, pred, target):
-        
-        quantile_losses = []
-        for i, q in enumerate(self.quantiles):
-            error = target - pred[..., i]
-            quantile_error = torch.max(
-                (q - 1) * error,
-                q * error
-            ).unsqueeze(-1)
-            quantile_losses.append(quantile_error)
-
-        quantile_losses = torch.cat(quantile_losses, dim = 2)
-        return quantile_losses
+from src.deep_learning.quantile_loss import QuantileLoss
 
 
 class LITransformer(L.LightningModule):
@@ -154,6 +20,9 @@ class LITransformer(L.LightningModule):
 
     # Initialize model
     def __init__(self, hyperparams_dict):
+    """
+    hyperparams_dict: Dictionary of hyperparameter values. See below for required hyperparameters.
+    """
         
         # Delegate function to parent class
         super().__init__() 
@@ -162,18 +31,18 @@ class LITransformer(L.LightningModule):
         self.save_hyperparameters(logger = False) 
 
         # Define hyperparameters
-        self.source_length = hyperparams_dict["source_length"] # Length of source sequence, L
-        self.target_length = hyperparams_dict["target_length"] # Length of target sequence, H
-        self.horizon_start = hyperparams_dict["horizon_start"] # Start of the forecast horizon relevant for loss computing
-        self.quantiles = hyperparams_dict["quantiles"] # Provide as list of floats: [0.025, 0.5, 0.975]
-        self.d_model = hyperparams_dict["d_model"] # Dimensionality of attention inputs, d
-        self.n_heads = hyperparams_dict["n_heads"] # N. of attention heads per multiattention block
-        self.n_encoders = hyperparams_dict["n_encoders"] # N. of encoder blocks
-        self.n_decoders = hyperparams_dict["n_decoders"] # N. of decoder blocks
-        self.d_feedforward = hyperparams_dict["d_feedforward"] # Dimensionality of feedforward networks
-        self.activation = hyperparams_dict["activation"] # Activation function for transformer FFNs
+        self.source_length = hyperparams_dict["source_length"]  # Length of source sequence, L
+        self.target_length = hyperparams_dict["target_length"]  # Length of target sequence, H
+        self.horizon_start = hyperparams_dict["horizon_start"]  # Start of the forecast horizon relevant for loss computing
+        self.quantiles = hyperparams_dict["quantiles"]  # Provide as list of floats: [0.025, 0.5, 0.975]
+        self.d_model = hyperparams_dict["d_model"]  # Dimensionality of attention inputs, d
+        self.n_heads = hyperparams_dict["n_heads"]  # N. of attention heads per multiattention block
+        self.n_encoders = hyperparams_dict["n_encoders"]  # N. of encoder blocks
+        self.n_decoders = hyperparams_dict["n_decoders"]  # N. of decoder blocks
+        self.d_feedforward = hyperparams_dict["d_feedforward"]  # Dimensionality of feedforward networks
+        self.activation = hyperparams_dict["activation"]  # Activation function for transformer FFNs
         self.learning_rate = hyperparams_dict["learning_rate"]
-        self.lr_decay = hyperparams_dict["lr_decay"]
+        self.lr_decay = hyperparams_dict["lr_decay"]  # Learning rate decay, 0.9 to 1 in practice, 1 = no decay
         self.dropout_rate = hyperparams_dict["dropout_rate"]
 
         # Define loss function: Quantile loss
@@ -412,4 +281,3 @@ class LITransformer(L.LightningModule):
           "scheduler": lr_scheduler
           }
         }
-
